move QgsCircularString tests from QgsGeometry
